ci(metrics): validate prometheus exposition and key series
