+++ services/radio-coord/src/main.rs
@@ -13,8 +13,8 @@ use common_config::{load, MsgBusConfig, ServiceConfig};
 use common_mdns::announce;
 use common_msgbus::{MessageBus, NatsBus, NatsConfig};
 use common_obs::{
-    encode_prometheus_metrics, handler_latency_seconds, http_requests_total, msgbus_publish_total,
-    ObsInit, PROMETHEUS_CONTENT_TYPE,
+    encode_prometheus_metrics, http_request_observe, msgbus_publish_total, ObsInit,
+    PROMETHEUS_CONTENT_TYPE,
 };
 use parking_lot::RwLock;
 use serde::{Deserialize, Serialize};
@@ -112,8 +112,7 @@ async fn track_http_metrics(req: Request<Body>, next: Next) -> Response {
     let latency = start.elapsed().as_secs_f64();
     let status = response.status().as_u16().to_string();
 
-    http_requests_total().inc(&[SERVICE_NAME, route.as_str(), status.as_str()], 1);
-    handler_latency_seconds().observe(&[SERVICE_NAME, route.as_str()], latency);
+    http_request_observe!(route.as_str(), status.as_str(), latency);
 
     response
 }
