+++ services/updater/Cargo.toml
@@ -9,3 +9,6 @@ common-config = { workspace = true }
 common-obs = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
+
+[dev-dependencies]
+reqwest = { version = "0.11", default-features = false, features = ["rustls-tls"] }
